Refactor: #241 modify home component
