+++ munetic_app/src/components/auth/Login.tsx
@@ -82,7 +82,6 @@ export default function Login() {
         ] = `Bearer ${accessToken}`;
         try {
           localStorage.setItem('user', JSON.stringify(login_id));
-          //localStorage.setItem('')
         } catch (e) {
           console.log(e, 'localStorage is not working');
         }
